Lock down faraday to < 1.0
Faraday 1.0 removed the deprecated Faraday::Error::ClientError in favor
of Faraday::ClientError.  The former is currently used in
follow_redirects middleware.  While this commit changes the code to use
the new Faraday::ClientError, which was introduced in faraday 0.9, we
still need to lock down to < 1.0 as there any many other failures as
well.